tracker: decode upstream stats as float64
tracker.gg ships numeric stats as JSON floats (e.g. 'value: 1067.0')
even for integer-valued fields like MMR. Decoding into an int field
fails with 'cannot unmarshal number 1067.0 into Go struct field
.data.segments.stats.rating.value of type int', producing a
'upstream error' response with no recoverable detail.

The hand-crafted fixture used JSON integers, so the unit tests never
exercised the float path. Bump the 3v3 rating + matchesPlayed in the
fixture to floats so the parser tests pin the real upstream shape.

Add probe_test.go (build-tagged 'probe', excluded from CI) that hits
real tracker.gg via the surf doer. Useful when you need to know
whether a regression is in our code or upstream behavior.

+++ backend/internal/tracker/parser.go
@@ -31,9 +31,11 @@ type upstreamSegment struct {
 	Metadata struct {
 		Name string `json:"name"`
 	} `json:"metadata"`
+	// tracker.gg ships numeric stats as JSON floats (e.g. "value":1067.0),
+	// even for integers like MMR. We decode as float64 and cast on output.
 	Stats struct {
 		Rating struct {
-			Value int `json:"value"`
+			Value float64 `json:"value"`
 		} `json:"rating"`
 		Tier struct {
 			Metadata struct {
@@ -46,7 +48,7 @@ type upstreamSegment struct {
 			} `json:"metadata"`
 		} `json:"division"`
 		MatchesPlayed struct {
-			Value int `json:"value"`
+			Value float64 `json:"value"`
 		} `json:"matchesPlayed"`
 	} `json:"stats"`
 }
@@ -78,10 +80,10 @@ func parseProfile(body []byte, platform, id string, now time.Time) (*Result, err
 			continue
 		}
 		out.Playlists[key] = Rating{
-			MMR:      seg.Stats.Rating.Value,
+			MMR:      int(seg.Stats.Rating.Value),
 			Tier:     seg.Stats.Tier.Metadata.Name,
 			Division: seg.Stats.Division.Metadata.Name,
-			Matches:  seg.Stats.MatchesPlayed.Value,
+			Matches:  int(seg.Stats.MatchesPlayed.Value),
 		}
 	}
 	return out, nil
